requirements.txt
torchvision
pillow
pdf2image
pytesseract
opencv-python-headless
rapidfuzz
numpy
torch
transformers
datasets
